+++ test/benchmark/step_benchmark_test.cc
@@ -74,7 +74,7 @@ static void run_step_benchmark(const mjModel* model, benchmark::State& state) {
 
 void ABSL_ATTRIBUTE_NO_TAIL_CALL BM_StepFlagPlugin(benchmark::State& state) {
   MujocoErrorTestGuard guard;
-  static mjModel* model = LoadModelFromPath("plugin/elasticity/flag.xml");
+  static mjModel* model = LoadModelFromPath("plugin/elasticity/flag_flex.xml");
   run_step_benchmark(model, state);
 }
 BENCHMARK(BM_StepFlagPlugin);